Disable electric-indent-local-mode, use newline-and-indent
Upon RET, the electric-indent stuff does an indent-line for the
original line, as well as an indent-line for the new line. The former
doesn't work well with hash-lang indenters. Although maybe I should
figure out why, it also seems reasonable simply to rebind RET to
newline-and-indent, which this commit does.