Improve handling of changed #lang
Only check for new #lang when a change position is <= the end of the
last lang spec we read.

When we have a changed lang, notify the front end. So now, there is a
'hash-lang notification of varieties 'lang or 'token.

The notification includes lang-info information like
paren-matches, quote-matches, and booleans for whether the lang
supplies a grouping-position, line-indenter, and range-indenter.

The idea here is to supply enough information to appropriately
configure things like the Emacs syntax-table, as well as
indent-line-function, indent-region-function, and
forward-sexp-function. That will be the next commit. The idea of this
commit is to get enough back end stuff sorted out for this that I can
focus on the Emacs end for awhile.