Save-point commit using token-tree%
This can be squashed out of existence; it's just a "save point"
reflecting me copying what color.rkt does, albeit simplified not to
have multiple "lexer-states".

However that basic approach to incremental lexing is not what we need.
This commit is to save that work just in case I realize I want to
refer to it. But then the next commit will throw it all away and start
over from scratch: Use our old approach modified as little as possible
to use token-tree% instead of interval-map.