Set plain syntax-table when not s-expr lang
The idea here -- which I'm not 100% confident about -- is that we can
use the standard syntax table for things like parens, strings, and
comments when the lang uses sexp navigation. There is no harm in doing
so, and it might allow more compatibility with other Emacs features
and packages. However when the lang does NOT use sexp nav, then we
assume nothing and clear the syntax table of such entries; instead we
rely entirely on the tokenization to apply 'syntax-table text
properties to portions of the buffer.