Fix problem closing unclosed @( in scribble/manual
With something like

    #lang scribble/manual
    text text
    text text

inserting this

    #lang scribble/manual
    @(if 1 2
    text text
    text text

causes all the text to be tokenized into symbol and white-space,
which is correct.

Adding a closing paren

    #lang scribble/manual
    @(if 1 2)
    text text
    text text

ought to cause stuff after the closing paren to be tokenized back to
text.

But it wasn't!

After a going on a detour worried that the hash-lang% do-update!
method was handling this change case incorrectly, I finally determined
it was due to Emacs blink-paren-face causing the change notification
to be discarded. WAT.

So mainly this commit just disables blink-paren-mode. But also:

Add a test of hash-lang% for this scenario.

Fine-tune some details about font-lock using jit-lock-mode, when it's
time to mark some span as needing to be refontified. AFAICT
flush-font-lock is not the correct thing to use. Instead remove the
'fontified property.

Strengthen racket--cmd-dispatch-response matching. Also this turned
out to be N/A, it was useful while I was investigating the Case of the
Disappearing Notification.