Reorganize/rename some things; only rebuild when lexer changes
Mostly this reorganizes and renames some things.

The only intentional substantive change is the handling of changed
lang-info values. If any change, we on-notify as before. However, we
only re-tokenize from scratch if the lexer changes. So e.g. if two
langs share a lexer but differ in (say) indenter that won't re-lex.