Exclude ./racket/hash-lang.rkt from raco test
This should have been part of commit dc6902b, where hash-lang.rkt
stopped doing dynamic-requires itself. Initially I'd set it up that
way to avoid just this kind of Makefile complication. But simplifying
it is part of keeping things on a trajectory toward it moving to
syntax-color-lib (and at that point the Makefile could become simpler
again). So accept the complication for now.